Added a TODO about adopting typed throws for AsyncSequence extensions once 2024 OS's are okay minimums.
Unfortunately the addition of the `Failure` associated type was only for Swift 6, in the 2024 OS releases (e.g. macOS 15).  I definitely don't want to drop support for pre-2024 OS versions just yet - maybe in a year or two.

In principle this should already be possible (by declaring two versions of the function, one for each side of the OS change where `Failure` was introduced) but it appears to be impossible to do this. 😕 https://forums.swift.org/t/conditionalising-function-declaration-on-os-version/72378